+++ src/main/java/de/maxhenkel/plane/entity/render/PlaneModel.java
@@ -3,74 +3,19 @@ package de.maxhenkel.plane.entity.render;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.math.Axis;
 import de.maxhenkel.corelib.client.obj.OBJModel;
-import de.maxhenkel.corelib.client.obj.OBJModelInstance;
-import de.maxhenkel.corelib.client.obj.OBJModelOptions;
-import de.maxhenkel.corelib.math.Rotation;
 import de.maxhenkel.plane.Main;
 import de.maxhenkel.plane.entity.EntityPlane;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
 import net.minecraft.resources.ResourceLocation;
-import org.joml.Vector3d;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
+import org.joml.Vector3f;
 
 public class PlaneModel extends AbstractPlaneModel<EntityPlane> {
 
-    private static final List<OBJModelInstance<EntityPlane>> MODELS = Arrays.asList(
-            new OBJModelInstance<>(
-                    new OBJModel(
-                            ResourceLocation.fromNamespaceAndPath(Main.MODID, "models/entity/wheel.obj")
-                    ),
-                    new OBJModelOptions<>(
-                            ResourceLocation.fromNamespaceAndPath(Main.MODID, "textures/entity/wheel.png"),
-                            new Vector3d(-10D / 16D, 2D / 16D, -17.5D / 16D),
-                            (plane, matrixStack, partialTicks) -> {
-                                matrixStack.scale(1F / 16F, 1F / 16F, 1F / 16F);
-                                matrixStack.mulPose(Axis.XP.rotationDegrees(-plane.getWheelRotation(partialTicks)));
-                            }
-                    )
-            ),
-            new OBJModelInstance<>(
-                    new OBJModel(
-                            ResourceLocation.fromNamespaceAndPath(Main.MODID, "models/entity/wheel.obj")
-                    ),
-                    new OBJModelOptions<>(
-                            ResourceLocation.fromNamespaceAndPath(Main.MODID, "textures/entity/wheel.png"),
-                            new Vector3d(10D / 16D, 2D / 16D, -17.5D / 16D),
-                            (plane, matrixStack, partialTicks) -> {
-                                matrixStack.scale(1F / 16F, 1F / 16F, 1F / 16F);
-                                matrixStack.mulPose(Axis.XP.rotationDegrees(-plane.getWheelRotation(partialTicks)));
-                            }
-                    )
-            ),
-            new OBJModelInstance<>(
-                    new OBJModel(
-                            ResourceLocation.fromNamespaceAndPath(Main.MODID, "models/entity/propeller.obj")
-                    ),
-                    new OBJModelOptions<>(
-                            ResourceLocation.withDefaultNamespace("textures/block/spruce_planks.png"),
-                            new Vector3d(0D / 16D, 16D / 16D, -29.5D / 16D),
-                            (plane, matrixStack, partialTicks) -> {
-                                matrixStack.scale(1F / 16F, 1F / 16F, 1F / 16F);
-                                matrixStack.mulPose(Axis.ZP.rotationDegrees(-plane.getPropellerRotation(partialTicks)));
-                            }
-                    )
-            )
-    );
-
-    private static final List<OBJModelInstance<EntityPlane>> OAK_MODEL = getPlaneModel(ResourceLocation.withDefaultNamespace("textures/block/oak_planks.png"));
-    private static final List<OBJModelInstance<EntityPlane>> DARK_OAK_MODEL = getPlaneModel(ResourceLocation.withDefaultNamespace("textures/block/dark_oak_planks.png"));
-    private static final List<OBJModelInstance<EntityPlane>> BIRCH_MODEL = getPlaneModel(ResourceLocation.withDefaultNamespace("textures/block/birch_planks.png"));
-    private static final List<OBJModelInstance<EntityPlane>> JUNGLE_MODEL = getPlaneModel(ResourceLocation.withDefaultNamespace("textures/block/jungle_planks.png"));
-    private static final List<OBJModelInstance<EntityPlane>> ACACIA_MODEL = getPlaneModel(ResourceLocation.withDefaultNamespace("textures/block/acacia_planks.png"));
-    private static final List<OBJModelInstance<EntityPlane>> SPRUCE_MODEL = getPlaneModel(ResourceLocation.withDefaultNamespace("textures/block/spruce_planks.png"));
-    private static final List<OBJModelInstance<EntityPlane>> WARPED_MODEL = getPlaneModel(ResourceLocation.withDefaultNamespace("textures/block/warped_planks.png"));
-    private static final List<OBJModelInstance<EntityPlane>> CRIMSON_MODEL = getPlaneModel(ResourceLocation.withDefaultNamespace("textures/block/crimson_planks.png"));
-    private static final List<OBJModelInstance<EntityPlane>> BAMBOO_MODEL = getPlaneModel(ResourceLocation.withDefaultNamespace("textures/block/bamboo_planks.png"));
-    private static final List<OBJModelInstance<EntityPlane>> CHERRY_MODEL = getPlaneModel(ResourceLocation.withDefaultNamespace("textures/block/cherry_planks.png"));
-    private static final List<OBJModelInstance<EntityPlane>> MANGROVE_MODEL = getPlaneModel(ResourceLocation.withDefaultNamespace("textures/block/mangrove_planks.png"));
+    private static final OBJModel PLANE_MODEL = new OBJModel(ResourceLocation.fromNamespaceAndPath(Main.MODID, "models/entity/plane.obj"));
+    private static final Vector3f BODY_OFFSET = new Vector3f(0F, 8F / 16F, 0F);
+    private static final Vector3f PROPELLER_OFFSET = new Vector3f(0F / 16F, 16F / 16F, -29.5F / 16F);
+    private static final Vector3f LEFT_WHEEL_OFFSET = new Vector3f(-10F / 16F, 2F / 16F, -17.5F / 16F);
+    private static final Vector3f RIGHT_WHEEL_OFFSET = new Vector3f(10F / 16F, 2F / 16F, -17.5F / 16F);
 
     public PlaneModel(EntityRendererProvider.Context renderManager) {
         super(renderManager);
@@ -88,12 +33,32 @@ public class PlaneModel extends AbstractPlaneModel<EntityPlane> {
     }
 
     @Override
-    public List<OBJModelInstance<EntityPlane>> getModels(EntityPlane entity) {
-        return getModelFromType(entity);
+    protected Vector3f getLeftWheelOffset(EntityPlane plane) {
+        return LEFT_WHEEL_OFFSET;
     }
 
+    @Override
+    protected Vector3f getRightWheelOffset(EntityPlane plane) {
+        return RIGHT_WHEEL_OFFSET;
+    }
 
-    private static List<OBJModelInstance<EntityPlane>> getModelFromType(EntityPlane plane) {
+    @Override
+    protected Vector3f getPropellerOffset(EntityPlane plane) {
+        return PROPELLER_OFFSET;
+    }
+
+    @Override
+    protected Vector3f getBodyOffset(EntityPlane plane) {
+        return BODY_OFFSET;
+    }
+
+    @Override
+    protected OBJModel getBodyModel(EntityPlane plane) {
+        return PLANE_MODEL;
+    }
+
+    @Override
+    protected ResourceLocation getBodyTexture(EntityPlane plane) {
         switch (plane.getPlaneType()) {
             default:
             case OAK:
@@ -119,22 +84,6 @@ public class PlaneModel extends AbstractPlaneModel<EntityPlane> {
             case MANGROVE:
                 return MANGROVE_MODEL;
         }
-    }
-
-    private static List<OBJModelInstance<EntityPlane>> getPlaneModel(ResourceLocation texture) {
-        List<OBJModelInstance<EntityPlane>> models = new ArrayList<>(MODELS);
-        models.add(new OBJModelInstance<>(
-                new OBJModel(
-                        ResourceLocation.fromNamespaceAndPath(Main.MODID, "models/entity/plane.obj")
-                ),
-                new OBJModelOptions<>(
-                        texture,
-                        new Vector3d(0D, 8D / 16D, 0D),
-                        new Rotation(180F, Axis.YP),
-                        (plane, matrixStack, partialTicks) -> matrixStack.scale(1F / 16F, 1F / 16F, 1F / 16F)
-                )
-        ));
-        return models;
     }
 
 }