+++ changelog.md
@@ -1,1 +1,3 @@
 - Improved damage when colliding horizontally
+- Optimized plane rendering
+- Improved plane rotation visuals